testapp/requirements.txt
-e ..
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
black>22.1.0